+++ src/main/java/org/matsim/freight/logistics/example/lsp/initialPlans/MyLSPScorer.java
@@ -3,7 +3,7 @@ package org.matsim.freight.logistics.example.lsp.initialPlans;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.matsim.freight.logistics.*;
-import org.matsim.freight.logistics.resourceImplementations.transshipmentHub.TransshipmentHubResource;
+import org.matsim.freight.logistics.resourceImplementations.TransshipmentHubResource;
 
 /**
  * A scorer for the LSP. It uses the scores of the - carriers: Take the carrier's score and add it
