+++ src/main/java/org/matsim/freight/logistics/example/lsp/multipleChains/ExampleGroceryDeliveryMultipleChains.java
@@ -27,7 +27,6 @@ import org.matsim.freight.carriers.controler.CarrierScoringFunctionFactory;
 import org.matsim.freight.carriers.controler.CarrierStrategyManager;
 import org.matsim.freight.logistics.*;
 import org.matsim.freight.logistics.resourceImplementations.ResourceImplementationUtils;
-import org.matsim.freight.logistics.resourceImplementations.transshipmentHub.TranshipmentHubUtils;
 import org.matsim.freight.logistics.shipment.LSPShipment;
 import org.matsim.vehicles.VehicleType;
 
@@ -242,13 +241,13 @@ final class ExampleGroceryDeliveryMultipleChains {
                 .build();
 
         LSPResourceScheduler hubScheduler =
-            TranshipmentHubUtils.TranshipmentHubSchedulerBuilder.newInstance()
+            ResourceImplementationUtils.TranshipmentHubSchedulerBuilder.newInstance()
                 .setCapacityNeedFixed(10)
                 .setCapacityNeedLinear(1)
                 .build();
 
         LSPResource hubResource =
-            TranshipmentHubUtils.TransshipmentHubBuilder.newInstance(
+            ResourceImplementationUtils.TransshipmentHubBuilder.newInstance(
                     Id.create("Hub", LSPResource.class), HUB_LINK_ID, scenario)
                 .setTransshipmentHubScheduler(hubScheduler)
                 .build();
