+++ src/main/java/org/matsim/freight/logistics/resourceImplementations/TransshipmentHubResource.java
@@ -1,4 +1,25 @@
 /*
+  *********************************************************************** *
+  * project: org.matsim.*
+  *                                                                         *
+  * *********************************************************************** *
+  *                                                                         *
+  * copyright       :  (C) 2024 by the members listed in the COPYING,       *
+  *                   LICENSE and WARRANTY file.                            *
+  * email           : info at matsim dot org                                *
+  *                                                                         *
+  * *********************************************************************** *
+  *                                                                         *
+  *   This program is free software; you can redistribute it and/or modify  *
+  *   it under the terms of the GNU General Public License as published by  *
+  *   the Free Software Foundation; either version 2 of the License, or     *
+  *   (at your option) any later version.                                   *
+  *   See also COPYING, LICENSE and WARRANTY file                           *
+  *                                                                         *
+  * ***********************************************************************
+ */
+
+/*
  *  *********************************************************************** *
  *  * project: org.matsim.*
  *  * *********************************************************************** *
@@ -18,7 +39,7 @@
  *  * ***********************************************************************
  */
 
-package org.matsim.freight.logistics.resourceImplementations.transshipmentHub;
+package org.matsim.freight.logistics.resourceImplementations;
 
 import java.util.Collection;
 import java.util.List;
@@ -29,6 +50,7 @@ import org.matsim.freight.logistics.LSPDataObject;
 import org.matsim.freight.logistics.LSPPlan;
 import org.matsim.freight.logistics.LSPResource;
 import org.matsim.freight.logistics.LogisticChainElement;
+import org.matsim.freight.logistics.resourceImplementations.ResourceImplementationUtils.TransshipmentHubBuilder;
 
 /**
  * {@link LSPResource} bei der die geplanten Tätigkeiten NICHT am Verkehr teilnehmen.
@@ -66,7 +88,7 @@ public class TransshipmentHubResource extends LSPDataObject<LSPResource> impleme
   }
 
   TransshipmentHubResource(
-      TranshipmentHubUtils.TransshipmentHubBuilder builder, Scenario scenario) {
+      TransshipmentHubBuilder builder, Scenario scenario) {
     super(builder.getId());
     this.locationLinkId = builder.getLocationLinkId();
     this.transshipmentHubScheduler = builder.getTransshipmentHubScheduler();
