+++ src/main/java/org/matsim/freight/logistics/resourceImplementations/TransshipmentHubScheduler.java
@@ -1,4 +1,25 @@
 /*
+  *********************************************************************** *
+  * project: org.matsim.*
+  *                                                                         *
+  * *********************************************************************** *
+  *                                                                         *
+  * copyright       :  (C) 2024 by the members listed in the COPYING,       *
+  *                   LICENSE and WARRANTY file.                            *
+  * email           : info at matsim dot org                                *
+  *                                                                         *
+  * *********************************************************************** *
+  *                                                                         *
+  *   This program is free software; you can redistribute it and/or modify  *
+  *   it under the terms of the GNU General Public License as published by  *
+  *   the Free Software Foundation; either version 2 of the License, or     *
+  *   (at your option) any later version.                                   *
+  *   See also COPYING, LICENSE and WARRANTY file                           *
+  *                                                                         *
+  * ***********************************************************************
+ */
+
+/*
  *  *********************************************************************** *
  *  * project: org.matsim.*
  *  * *********************************************************************** *
@@ -18,7 +39,7 @@
  *  * ***********************************************************************
  */
 
-package org.matsim.freight.logistics.resourceImplementations.transshipmentHub;
+package org.matsim.freight.logistics.resourceImplementations;
 
 import java.util.ArrayList;
 import org.apache.logging.log4j.LogManager;
@@ -28,6 +49,7 @@ import org.matsim.freight.logistics.LSPResource;
 import org.matsim.freight.logistics.LSPResourceScheduler;
 import org.matsim.freight.logistics.LogisticChainElement;
 import org.matsim.freight.logistics.LspShipmentWithTime;
+import org.matsim.freight.logistics.resourceImplementations.ResourceImplementationUtils.TranshipmentHubSchedulerBuilder;
 import org.matsim.freight.logistics.shipment.ShipmentPlan;
 import org.matsim.freight.logistics.shipment.ShipmentPlanElement;
 import org.matsim.freight.logistics.shipment.ShipmentUtils;
@@ -40,7 +62,7 @@ import org.matsim.freight.logistics.shipment.ShipmentUtils;
   private TransshipmentHubResource transshipmentHubResource;
   private TransshipmentHubTourEndEventHandler eventHandler;
 
-  TransshipmentHubScheduler(TranshipmentHubUtils.TranshipmentHubSchedulerBuilder builder) {
+  TransshipmentHubScheduler(TranshipmentHubSchedulerBuilder builder) {
     this.lspShipmentsWithTime = new ArrayList<>();
     this.capacityNeedLinear = builder.getCapacityNeedLinear();
     this.capacityNeedFixed = builder.getCapacityNeedFixed();
